refactor: create get random episode use case
